fix: add missing auth token validation in delete-photo and post-toggle-photo-like actions

+++ client/src/actions/photos/delete-photo.ts
@@ -3,13 +3,19 @@
 import { ApiResponse } from "@/interfaces/actions"
 import ApiError from "@/utils/api-error"
 import { PHOTO_DELETE_PHOTO } from "@/utils/api-urls"
+import { cookies } from "next/headers"
 
 export default async function deletePhoto(photoId: string) {
     const URL = PHOTO_DELETE_PHOTO(photoId)
 
     try {
+        const access_token = cookies().get("access_token")?.value
+        if (!access_token)
+            throw new Error("You must be logged in to delete a photo.")
+
         const response = await fetch(URL, {
             method: "DELETE",
+            headers: { Authorization: "Bearer " + access_token },
             next: { revalidate: 60, tags: ["photos"] }
         })
